feat: 답변 채택 기능 구현

+++ src/main/java/com/lawnroad/board/controller/LawyerCommentController.java
@@ -2,6 +2,8 @@ package com.lawnroad.board.controller;
 
 import com.lawnroad.board.dto.*;
 import com.lawnroad.board.service.CommentService;
+import com.lawnroad.common.util.JwtTokenUtil;
+import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -17,10 +19,19 @@ import java.util.List;
 public class LawyerCommentController {
 
     private final CommentService commentService;
+    private final JwtTokenUtil jwtUtil;
 
     // 답변 등록
     @PostMapping
-    public ResponseEntity<Void> registerComment(@RequestBody CommentRegisterDto dto) {
+    public ResponseEntity<Void> registerComment(@RequestBody CommentRegisterDto dto,
+                                                @RequestHeader("Authorization") String authHeader) {
+        // JWT에서 userNo 추출
+        String token = authHeader.replace("Bearer ", "");
+        Claims claims = jwtUtil.parseToken(token);
+        Long userNo = claims.get("no", Long.class);
+
+        dto.setUserNo(userNo);
+
         commentService.registerComment(dto);
         return ResponseEntity.status(201).build();
     }
@@ -34,29 +45,51 @@ public class LawyerCommentController {
     @GetMapping("/answers")
     public ResponseEntity<Page<MyCommentResponseDto>> getMyAnswers(
             @RequestParam(defaultValue = "1") int page,
-            @RequestParam(defaultValue = "10") int size
+            @RequestParam(defaultValue = "10") int size,
+            @RequestHeader("Authorization") String authHeader
     ) {
-        Long userNo = 32L; //TODO 나중 하드코딩 바꿔야됨
-//        System.out.println(page);
-//        System.out.println(size);
+        String token = authHeader.replace("Bearer ", "");
+        Claims claims = jwtUtil.parseToken(token);
+        Long userNo = claims.get("no", Long.class);
+
         Page<MyCommentResponseDto> result = commentService.getMyComments(userNo, page, size);
         return ResponseEntity.ok(result);
     }
     // 특정 댓글 상세 조회
     @GetMapping("/detail/{commentId}")
     public ResponseEntity<CommentDetailDto> getCommentDetail(@PathVariable Long commentId) {
-        Long userNo = 32L; // 임시로 하드코딩
-        return ResponseEntity.ok(commentService.findById(commentId, userNo));
+
+        return ResponseEntity.ok(commentService.findById(commentId));
     }
 
     @PutMapping("/{commentId}")
-    public ResponseEntity<Void> updateComment(@PathVariable Long commentId, @RequestBody CommentUpdateDto dto) {
+    public ResponseEntity<Void> updateComment(@PathVariable Long commentId, @RequestBody CommentUpdateDto dto, @RequestHeader("Authorization") String authHeader) {
+        String token = authHeader.replace("Bearer ", "");
+        Claims claims = jwtUtil.parseToken(token);
+        Long userNo = claims.get("no", Long.class);
+
+        // 검증 로직 (작성자 본인 확인)
+        CommentDetailDto comment = commentService.findById(commentId);
+        if (!comment.getUserNo().equals(userNo)) {
+            return ResponseEntity.status(403).build();
+        }
+
+        dto.setUserNo(userNo); // 보안상 다시 주입
         commentService.updateComment(commentId, dto);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/{commentId}")
-    public ResponseEntity<Void> deleteComment(@PathVariable Long commentId) {
+    public ResponseEntity<Void> deleteComment(@PathVariable Long commentId, @RequestHeader("Authorization") String authHeader) {
+        String token = authHeader.replace("Bearer ", "");
+        Claims claims = jwtUtil.parseToken(token);
+        Long userNo = claims.get("no", Long.class);
+
+        CommentDetailDto comment = commentService.findById(commentId);
+        if (!comment.getUserNo().equals(userNo)) {
+            return ResponseEntity.status(403).build(); // 권한 없음
+        }
+
         commentService.deleteComment(commentId);
         return ResponseEntity.noContent().build();
     }
